requirements.txt
absl-py==2.3.1
astunparse==1.6.3
bcrypt==5.0.0
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
docutils==0.22.4
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.2
Flask-SocketIO==5.6.0
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.13.0
Kivy==2.3.1
Kivy-Garden==0.1.5
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.0
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.2
Pygments==2.19.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-engineio==4.13.0
python-socketio==5.16.0
pytz==2025.2
requests==2.32.5
rich==14.2.0
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.5
wheel==0.45.1
wrapt==2.0.1
wsproto==1.3.2
